requirements.txt
# requirements.txt
# Dependencies for llm-telemetry-agent
# Generated: 2026-04-24
# Python 3.x
#
# Install with: pip install -r requirements.txt

fastapi==0.136.0
uvicorn==0.45.0
httpx==0.28.1
opentelemetry-api==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-instrumentation-fastapi==0.62b0
opentelemetry-instrumentation-httpx==0.62b0
